proyectos/01-chatbot-legal-basico/requirements.txt
# Chatbot Legal Básico - Dependencias
# Proyecto 01 - RAG MVP Legal

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-aws>=0.2.0

# AWS Bedrock
boto3>=1.34.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing (opcional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
